Fix TF2 3D Unet to standard model garden recommended style.
PiperOrigin-RevId: 306752053
